+++ src/utils/features.ts
@@ -9,6 +9,7 @@ const supportedFeatures = {
   advancedQueryRedirectConditions: { minVersion: '4.5.0' },
   desktopDeviceTypes: { minVersion: '4.5.0' },
   filterShortUrlsByExcludedTags: { minVersion: '4.6.0' },
+  filterVisitsByDomain: { minVersion: '4.6.0' },
 } as const satisfies Record<string, Versions>;
 
 Object.freeze(supportedFeatures);
@@ -27,6 +28,7 @@ const getFeaturesForVersion = (serverVersion: SemVerOrLatest): Record<Feature, b
   advancedQueryRedirectConditions: isFeatureEnabledForVersion('advancedQueryRedirectConditions', serverVersion),
   desktopDeviceTypes: isFeatureEnabledForVersion('advancedQueryRedirectConditions', serverVersion),
   filterShortUrlsByExcludedTags: isFeatureEnabledForVersion('filterShortUrlsByExcludedTags', serverVersion),
+  filterVisitsByDomain: isFeatureEnabledForVersion('filterVisitsByDomain', serverVersion),
 });
 
 const FeaturesContext = createContext(getFeaturesForVersion('0.0.0'));